pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiida-adsorption-wf"  # Choose a suitable project name
version = "0.1.0"
description = "A librairies of workflows to study adsorption in materials screening."
authors = [
    { name = "Arthur Hardiagon", email = "arthur.hardiagon@chimieparistech.psl.eu" }
]
dependencies = [
    "pytest",
    "aiida-raspa",
    "jupyterlab",
    "ipykernel"
  ]
[project.scripts]
eqeq_raspa_screening = "aiida_adsorption_wf.workflows.eqeq_raspa_screening:main"
analyze_workflow    = "aiida_adsorption_wf.utils.analyze_workflow:main"
plot_cpu_time        = "aiida_adsorption_wf.utils.plot_cpu_time:main"
check_and_convert_filenames        = "aiida_adsorption_wf.utils.check_and_convert_filenames:main"
calcjob2path        = "aiida_adsorption_wf.utils.calcjob2path:main"